third_party/WebKit/LayoutTests/external/wpt/payment-handler/payment-handler-exposed.https.html

<!DOCTYPE html>
<meta charset=utf-8>
<link rel="help" href="https://w3c.github.io/payment-handler/">
<title>PaymentInstruments.set() activates the associated service worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script>

async function receiveDataFromSW(serviceWorker) {
  return new Promise((resolve) => {
    const channel = new MessageChannel();
    channel.port1.addEventListener('message', (e) => {
      resolve(e.data);
    });

    serviceWorker.postMessage({port: channel.port2}, [channel.port2]);
    channel.port1.start();
  });
}

promise_test(async test => {
  const registration = await service_worker_unregister_and_register(
      test, 'resources/sw-payment-handler.js', 'resources/');
  await wait_for_state(test, registration.installing, 'activated');

  // Exposed Window
  assert_true('paymentManager' in registration);
  assert_true('userHint' in registration.paymentManager);
  assert_true('instruments' in registration.paymentManager);
  assert_true(await receiveDataFromSW(registration.active));
});

</script>
